Refactor
 * Strict typing
 * Simplify project setup
 * Add types for easier testing
